Major refactor to add examples for FHIR resources.
Moved around the CustomSkeletonAPI to be the
CustomSkeletonRestController.  Added classes and methods to follow the
OpenEMR pattern of FHIR naming conventions with Service class and
Controller class.  Put in basic example for extending the FHIR api using
fhir resouces and adding in a custom fhir resource as an example to give
people and idea of what they can do with the FHIR api.